Rename Hope to hope in packages
Renamed Hope to hope to comply with package naming conventions.
