--- src/main/java/Hope/commands/AddFixedDurationTaskCommand.java
@@ -1,56 +1,0 @@
-package Hope.commands;
-
-import Hope.storage.TaskStorage;
-import Hope.storage.ToDoList;
-import Hope.tasks.DeadlineTask;
-import Hope.tasks.FixedDurationTask;
-
-public class AddFixedDurationTaskCommand implements Command {
-    private ToDoList toDoList;
-    private TaskStorage taskStorage;
-
-    /**
-     * Creates a {@link AddDeadlineTaskCommand} with the specified to-do list and task storage
-     *
-     * @param toDoList the {@link ToDoList} that {@link DeadlineTask} will be added to
-     * @param taskStorage the {@link TaskStorage} that will be updated
-     */
-    public AddFixedDurationTaskCommand(ToDoList toDoList, TaskStorage taskStorage) {
-        this.toDoList = toDoList;
-        this.taskStorage = taskStorage;
-    }
-
-    /**
-     * Executes the command to add a deadline task based on the provided input and return a string.
-     * The input string is expected to contain a task description and a deadline, separated by "/by".
-     * If the input is invalid (e.g., missing description or deadline), an error message is printed,
-     * and no task is added. Otherwise, a {@link DeadlineTask} is created, added to the to-do list,
-     * storage updated, and a confirmation message printed along with the updated task count.
-     *
-     * @param o the input expected to be a {@code String} containing the task description
-     *          and deadline in the format "description /by deadline"
-     */
-    @Override
-    public String execute(Object o) {
-        String input = (String) o;
-        if (input.isEmpty()) {
-            return ("Thou hast overlooked the noble task of bestowing a worthy description upon this "
-                    + "endeavor!")
-                    + ("(Looks like you forgot to input a description and a deadline! Try again)\n");
-        }
-        String[] info = input.split("/needs");
-        if (info.length != 2) {
-            return "Verily, thou hast erred in thy response; endeavor once more, brave soul!"
-                    + "(Incorrect input, please try again)\n";
-        }
-        FixedDurationTask temp = new FixedDurationTask(info[0].trim(), info[1].trim());
-        toDoList.add(temp);
-        taskStorage.append(temp);
-        return "Behold, this quest hath been entrusted!\n"
-                + temp.toString()
-                + "\nLo! Thou art now bestowed with "
-                + toDoList.size()
-                + " noble quests upon thy parchment of duties.\n"
-                + "(You now have " + toDoList.size() + " tasks in the list)\n";
-    }
-}
