--- src/main/java/Hope/gui/GuiLauncher.java
@@ -1,12 +1,0 @@
-package Hope.gui;
-
-import javafx.application.Application;
-
-/**
- * A launcher class to workaround classpath issues.
- */
-public class GuiLauncher {
-    public static void main(String[] args) {
-        Application.launch(Gui.class, args);
-    }
-}
