+++ src/main/java/hope/gui/GuiLauncher.java
@@ -1,0 +1,12 @@
+package hope.gui;
+
+import javafx.application.Application;
+
+/**
+ * A launcher class to workaround classpath issues.
+ */
+public class GuiLauncher {
+    public static void main(String[] args) {
+        Application.launch(Gui.class, args);
+    }
+}
