+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Greetings, fair wanderer! I am Hope, your humble companion
+Greetings, fair wanderer! I am hope, your humble companion
 How might I assist thee on this grand quest?
 
 Behold, this quest hath been entrusted!
